backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
Flask-JWT-Extended==4.6.0
Flask-Talisman==1.1.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
bleach==6.1.0
bcrypt==4.1.2
cryptography==41.0.7
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow==10.2.0
Flask-SocketIO==5.3.6
python-socketio==5.10.0
eventlet==0.35.1
redis==5.0.1
celery==5.3.6
Flask-Mail==0.9.1
WeasyPrint==60.2
anthropic==0.23.0
marshmallow==3.20.1
email-validator==2.1.0
cloudinary==1.36.0
geoip2==4.7.0
requests==2.31.0
python-dotenv==1.0.0
shortuuid==1.0.11
